pyproject.toml
[project]
name = "mirror"
version = "0.1.0"
description = "Media asset management and publishing workflow engine"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "boto3",
  "botocore",
  "dateparser",
  "ffmpeg-python",
  "imagehash",
  "jsonschema",
  "levenshtein",
  "numpy",
  "opencv-python",
  "Pillow>=12.0.0",
  "python-dotenv",
  "PyYAML",
  "requests",
  "sqlean.py",
  "xmltodict",
  "markdown",
  "tomli-w",
  "typeguard",
  "feedgen",
  "psutil",
  "zahir2",
  "bookman",
  "tertius",
  "orbis",
  "textual>=8.2.3",
  "rich-pixels>=3.0.1",
  "textual-image>=0.11.0",
]

[project.optional-dependencies]
labeller = [
  "google-genai",
]
dev = [
  "vulture",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "BLE", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/mirror/commons/__init__.py" = ["F403", "F405"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
mirror = "mirror.cli:main"
[tool.setuptools]
package-dir = { "" = "src" }


[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
zahir2 = { path = "/home/rg/Code/zahir2", editable = true }
tertius = { path = "/home/rg/Code/tertius", editable = true }
orbis = { path = "/home/rg/Code/orbis", editable = true }
bookman = { path = "/home/rg/Code/sand", editable = true }
